requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
numpy
python-multipart
gunicorn
